fix(JSX) don't reload the transformer in dev; separate JSXTranformer test
