+++ lib/react/jsx.rb
@@ -6,24 +6,18 @@ require 'rails'
 
 module React
   module JSX
-    mattr_accessor :transform_options, :transformer_class
+    DEFAULT_TRANSFORMER = BabelTransformer
+    mattr_accessor :transform_options, :transformer_class, :transformer
 
     # You can assign `React::JSX.transformer_class = `
     # to provide your own transformer. It must implement:
     # - #initialize(options)
     # - #transform(code) => new code
-    self.transformer_class = BabelTransformer
+    self.transformer_class = DEFAULT_TRANSFORMER
 
     def self.transform(code)
-      transformer.transform(code)
-    end
-
-    def self.transformer
-      # lazily loaded during first request and reloaded every time when in dev or test
-      if @transformer.nil? || !::Rails.env.production?
-        @transformer = transformer_class.new(transform_options)
-      end
-      @transformer
+      self.transformer ||= transformer_class.new(transform_options)
+      self.transformer.transform(code)
     end
   end
 end
